requirements-dev.txt
django-stubs==5.0.4
